requirements.txt
shiny
pandas
numpy
meteostat
shinywidgets
plotly
python-dotenv
scipy
statsmodels
requests
datetime
libsass
